Add vim to images (#1186)
* add vim to images

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

We've found that having vim available on images is useful when needing
to debug. So this PR adds it to the ray-node, gateway, and repository images.

* drop vim from ray node

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

This adds a bit of size, and we run a LOT of ray nodes. The ray image
also allows for `sudo apt install vim`, so we don't need to preinstall
it.