implement allowlist framework for dependencies
Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

Allow operators to specify an allowlist of dependencies and allowed
versions.

The allowlist is stored in a config file.

A sample allowlist might look like:

    allowlist = { "wheel": ["0.44.0", "0.43.2"] }

which would imply that the wheel package is allowed, but only versions
0.43.2 and 0.44.0 .

For this PR, the allowlist is empty, which means all dependencies are
allowed, and it is stored locally. In a future PR, this should move
into a Kubernetes configmap so that it can be updated without having
to rebuild the container.

Each dependency can optionally specify a list of allowed versions. If
the list is empty, then all versions of the dependency are allowed.

An example:

    allowlist = { "wheel": [] }

This could also be adapted to store a minimum allowed version instead
of listing all allowed versions, depending on requirements.