add dependency parsing grammar for gateway allowlist
Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>

This PR adds the dependency specification grammar from PEP
508 (https://peps.python.org/pep-0508/) to use when validating
dependencies via the allowlist.

As a temporary measure, it also simplifies the check to only consider
the dependency and not a specific version. As such, it has dropped the
test for related to dependency version.

Future work should be done to convert the allowlist file from JSON to
the standard requirements.txt format, which would allow for easier
version checking.