Upgrade to qiskit-sphinx-theme 2.0.0 (#1481)
* Upgrade to qiskit-sphinx-theme 2.0.0

This upgrades to qiskit-sphinx-theme 2.0.0.  It includes the
Qiskit logo because the logo is no longer included in the theme.

qiskit-sphinx-theme 2.0 supports dark mode as a new feature, so
both light and dark versions of the logo are included.

* Crop empty space from logos

* Transparent background

* Make foreground color be #d0d0d0 to match theme's text

* Switch to svg version of logo

This is a cropped version of the old Qiskit Ecosystem logo,
taken from
https://github.com/Qiskit/qiskit_sphinx_theme/blob/f968969f2f503efbcfe611bb11239157de427ad3/src/qiskit_sphinx_theme/theme/qiskit-sphinx-theme/static/images/ecosystem-logo.svg

I created the dark mode version by hand-editing the resulting SVG
file to use the color #d0d0d0, which matches the qiskit-sphinx-theme
dark mode text color.